refactor(convert): extract shadow from extract_block_style
